backend/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Pydantic
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Auth
PyJWT[crypto]==2.9.0
cryptography>=42.0.0
bcrypt==4.2.1
python-multipart==0.0.20

# HTTP Client
httpx==0.28.1

# Task Queue
celery[redis]==5.4.0
redis==5.2.1
flower==2.0.1

# Utils
python-dotenv==1.0.1


